Add mode sql; table:prefectures

+++ structs/json.go
@@ -1,12 +1,12 @@
 package structs
 
 type Prefecture struct {
-	PrefCode int8 `json:"prefCode"`
+	PrefCode int    `json:"prefCode"`
 	PrefName string `json:"prefName"`
 }
 
 type Prefectures struct {
-	Message *string `json:"message"`
-	StatusCode string `json:"statusCode"`
-	Result []Prefecture `json:"result"`
+	Message    *string      `json:"message"`
+	StatusCode string       `json:"statusCode"`
+	Result     []Prefecture `json:"result"`
 }
